issue #147
